Move all page-type components to containers directory.
