+++ frontend/src/app/containers/HomePageWithAddon.js
