+++ frontend/src/app/containers/LoadingPage.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
-import Loading from './Loading';
+import Loading from '../components/Loading';
+
 
 export default function LoadingPage() {
   return (
