+++ frontend/src/app/containers/NotFoundPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import View from './View';
+import View from '../components/View';
 
 
 export default class NotFoundPage extends React.Component {
