+++ frontend/src/app/containers/RestartPage.js
