+++ frontend/src/app/containers/SharePage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import View from './View';
+import View from '../components/View';
 
 
 export default class SharePage extends React.Component {
