+++ frontend/test/app/containers/NotFoundPage-test.js
@@ -2,9 +2,10 @@ import React from 'react';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 
-import NotFoundPage from '../../../src/app/components/NotFoundPage';
+import NotFoundPage from '../../../src/app/containers/NotFoundPage';
 
-describe('app/components/NotFoundPage', () => {
+
+describe('app/containers/NotFoundPage', () => {
   it('should render notFoundHeader string', () => {
     expect(shallow(<NotFoundPage />)
       // HACK: .find('[data-l10n-id="errorMessage"]') seems not to work
