+++ frontend/test/app/containers/SharePage-test.js
@@ -3,9 +3,10 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 import { shallow } from 'enzyme';
 
-import SharePage from '../../../src/app/components/SharePage';
+import SharePage from '../../../src/app/containers/SharePage';
 
-describe('app/components/SharePage', () => {
+
+describe('app/containers/SharePage', () => {
   const linknames = ['facebook', 'twitter', 'email', 'copy'];
 
   let props, subject;
